\name{dataBiblio}
\alias{dataBiblio}
\docType{data}
\title{dataBiblio (data)}
\description{
A total of 386 abstracts on Systemic Lupus Erythematosus published between January 2000 and December 2012.

}
\usage{data("dataBiblio")}
\format{
Data frame with 386 abstracts. The rows represent the document (abstracts), the columns represent the variables which describe the abstracts (Title, Abstract, Year, Author, Journal).

    }

\keyword{datasets}