\name{dataOpen.question}

\alias{dataOpen.question}

\title{dataOpen.question (data)}

\description{
Data issued from a survey. Three questions are included in the questionnaire:  
"What do you wish for your family?", "What else?" and "What does culture mean for you?". 
These questions require free answers.
}

\usage{data("dataOpen.question")}


\format{
Data frame with 100 rows and 8 columns. The rows represent the documents (respondents) and the columns the variables. The first five columns correspond to the individuals characteristics (Gender,Age_interval,Gender_Age,Gender_educ,ageeduc) and the last three columns correspond to the open-ended questions.
}


\keyword{datasets}
