\name{dataSpeech}
\alias{dataSpeech}
\docType{data}
\title{dataSpeech (data)}

\description{

Speech "I Have a Dream", by Martin Luther King, Jr. delivered on 28 August 1963, at the Lincoln Memorial, Washington D.C.


}

\usage{data("dataSpeech")}

\format{
A data frame with one textual column.
}


\keyword{datasets}
