\name{TextoMineR-package}
\alias{TextoMineR}
\docType{package}

\title{Textual Statistics}
\description{
Textual Statistics.
}
\details{
\tabular{ll}{
Package: \tab TextoMineR\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-12-10\cr
License: GPL (>=2.0)
}

}
\author{
Monica Becue-Bertaut, Daria M. Hernandez, Belchin Kostov, Josep A. Sanchez
}

